requirements.txt
# requirements.txt

aiogram==3.7.0
google-generativeai==0.5.4
pydub==0.25.1
openai-whisper==20231117
aiohttp==3.9.5
python-dotenv==1.0.1
PyYAML==6.0.1